EMG/requirements.txt
numpy
pandas
tensorflow
scikit-learn
pyserial
statsmodels